add an example

+++ src/cz1/hmm/tools/AssemblyError.java
@@ -131,7 +131,7 @@ public class AssemblyError extends RFUtils {
 		}
 		
 		if(myArgsEngine.getBoolean("-phi")) {
-			skew_phi = Integer.parseInt(myArgsEngine.getString("-phi"));
+			skew_phi = Double.parseDouble(myArgsEngine.getString("-phi"));
 		}
 		
 		if(myArgsEngine.getBoolean("-nd")) {
