+++ src/cz1/hmm/tools/HaplotypeAnalysis.java
@@ -78,7 +78,7 @@ public class HaplotypeAnalysis extends RFUtils {
 		}
 		
 		if(myArgsEngine.getBoolean("-phi")) {
-			skew_phi = Integer.parseInt(myArgsEngine.getString("-phi"));
+			skew_phi = Double.parseDouble(myArgsEngine.getString("-phi"));
 		}
 		
 		this.best_n = 1;
