+++ src/cz1/hmm/tools/SinglePointAnalysis.java
@@ -109,7 +109,7 @@ public class SinglePointAnalysis extends RFUtils {
 		}
 		
 		if(myArgsEngine.getBoolean("-phi")) {
-			skew_phi = Integer.parseInt(myArgsEngine.getString("-phi"));
+			skew_phi = Double.parseDouble(myArgsEngine.getString("-phi"));
 		}
 		
 		if(myArgsEngine.getBoolean("-nd")) {
